Kill JsonCpp with fire.
Whoa! It's really a big commit.
In Cocaine Core, JsonCpp is used only to read a config when the core is initializing and Json::Value is used to store json-like data. But this library is big, it's slow, and it must be installed with the core. So i've removed it. Now cocaine::dynamic_t is used instead of Json::Value. RapidJson is used to parse a config, because it's header-only.